Provide possibility to register back channels with the cache manager
The back channels work as an opportunistic publish-subscribe mechanism, allowing the cache manager to send information to its clients. Clients can register a pipe with a channel id, an arbitrary string that will be Md5 hashed. The cache manager broadcasts to all registered channels in non-blocking mode.  On errors, the message is lost.  The cache manager will automatically unregister a back channel if the pipe is broken (or on any other error than EAGAIN).

This patch also fixes that broken return pipes kill the cache manager.

This is the foundation for notifications about a high gauge of pinned catalogs.