First revision of grafting interface.
For any given file in the repository ("regular" files only) named $FOO,
if there is an equivalent file .cvmfsgraft-$FOO in the same directory,
instead of processing the file normally, CVMFS will parse the .cvmfsgraft
file in order to get the checksum and filesize.

It is then the repo maintainer's responsibility to make the file available
in the repository.

Chunked files are not currently supported.