First draft of cvmfs_cp
cvmfs_cp is a small cp-like CLI for CVMFS.  It allows you to copy
files in and out of a CVMFS repo, regardless of how the host is
setup.

Beyond immediate value in testing independently of the host config,
this would allow "light" users of CVMFS to utilize CVMFS's capabilities
even on sites where it is not mounted.