Fill in repository options as best possible.
To the best of our ability, auto-fill repository and global options
to reduce the number of flags users need to pass:
- Take proxy from http_proxy environment variable.
- Create and cleanup a cache directory if none were specified.
- Add default URLs for repos (based on those shipped by CVMFS source).
- If instructed, add default signing keys (based on those shipped by
  CVMFS source).

The last bullet point is a bit controversial - it makes it difficult
to revoke keys if they are builtin to the binaries.  However, the
other option we give to users is to ignore the signing altogether;
I view this as the lesser of two evils.

Overall, the options are:
1) Provide your own public key files (suggested).
2) Use built-in, possibly revoked keys.
3) Ignore signing.
If need be, we can remove (2) before release.