Have FUSE process report the currently-used pubkey.
Allow the signature manager to track the pubkey (either from list
of keys or the certificate's pubkey) last used to verify data, and
expose this information as an extended attribute.

With this, we can deduce all the critical information (host and
pubkey) necessary to access a mounted repo via libcvmfs (where
the access may happen on a different worker node).