Require explicit VO names.
In case two VOs (purposely or accidentally) start using overlapping
group names, require the VOMS authz to explicitly name which VO is
being used.

For example, the LIGO VO issues attributes of the form /ligo; it
appears it's not safe to assume the VO name is part of the group
name.