Cache SSL context info.
Have the download job cache a pointer the proxy data used by the
SSL context.  Since the SSL context will need to be re-created
after a redirect, we cannot give the memory to the context object
itself - we must manage it explicitly.

Without this, the SSL context created for the redirection target
will drop the user proxy and end up unauthenticated.