Fix CURL flags
When a file descriptor is ready for _writing_ (POLLOUT), it was incorrectly being marked as ready for _reading_ in `libcurl`.  This caused busy wait loops in the download thread, where `libcurl` would try to read from a specified descriptor and keep getting EAGAIN.

In testing, this gave a respectable 10% speedup in downloads.