Avoid busy-wait when downloading files
CVMFS currently wakes up once a millisecond to check if libcurl has any ongoing transfers which have timed out.  As the granularity of our timeouts is in seconds, this is overkill: we can back off and save system CPU time.

libcurl actually provides a callback interface where it will inform CVMFS when the next upcoming library timeout will occur.  However, due to the proximity of the release, I simply increase the granularity.