Avoid use-after-free in OpenSSL.
In the case where a user with a X509 proxy accesses a file
then deletes the proxy and tries to download a second file, the
value of CURLOPT_SSL_CTX_DATA would not be overwritten when the
curl handle object was reused.  The X509 chain in-memory *would*
be deleted, however.

On the second use, the OpenSSL context would be given the invalid
pointer to the X509 chain by libcurl, resulting in a segfault.