Disable VOMS checks.
Unfortnuately, libvomsapi doesn't do all the validation we need
for this to function right.  For now, we disable VOMS checks and
always assume they fail - we'll fix this up in the next devel series.

This does allow HTTPS to be used for URLs as we blindly pass the
client credential to the remote host and don't need to validate it
in the client.