Initial implementation of a tiered cache.
This implementation provides a hierarchical write-through cache.
It allows a site to setup a local disk cache in addition to a site-
wide alien cache.

I envision this to be useful in two cases:
1) The performance of the alien cache (or, in the future, object
   storage) is better than a HTTP proxy setup because the underlying
   storage system allows several storage servers to peer.
2) When the repository is a secure CVMFS repo, as secure repos cannot
   be used with HTTP proxies.

No tests provided at this point - I just want to make sure this is
the correct approach, coding-wise.