Add support for handling special files when using OverlayFS
This commit fixes the problem with publishing a transaction
when there are special files added to the repository
and OverlayFS is used as a UnionFS driver.

Whiteout files in OverlayFS are represented as character devices
with both major and minor numbers 0. OverlayFS driver doesn't allow
users to create such files.

Both block and character devices will be ignored with a warning message
printed out.

A new SyncItemType is introduced: kBlockDevice

SyncItem has 4 new methods:
 - GetMajor
 - GetMinor
 - IsBlockDevice
 - IsSpecialFile

Handling special files may change in future in the following ways:
 - fail by default, enable explicitly in config to ignore with warning
 - add special files to the repo