Fix off-by-one error for chunk size.
Previously, a new chunk would be emitted after N+(buffer size)
bytes when the user requested N bytes for chunk size.  So, if the
user wanted a 24MB chunk, CVMFS would create 24MB+16KB chunks.

This corrects the issue; 24MB chunks are created when 24MB chunks
are requested for grafts.