Tiered cache txn should succeed if upper tier txn does.
Previously, the tiered cache counted a transaction as successful
if *both* caches had a successful transaction.  Thus, the reliability
of the tiered cache was the _AND_ of both caches -- i.e., slightly
decreased.

This changes the success metric so the transaction commit succeeds
if and only if the upper cache transaction works (the upper cache
*must* work because we may have handed out a FD to the cache via
OpenFromTxn already).