Add option to create catalog during ingestion.
This pull request add a flag (-c, short for --catalog) that allow to
automagically add a (sub-)catalog when ingesting a tar file.
The catalog is added in the base dir, where the tar is "uncompressed".

I am not so sure about this implementation.
It is not a very clean design, we borrow a method from a nice interface
(AddCatalogMarker from CatalogBalancer), similarly when we are actually
adding the catalog (in SyncMediator::Commit) we are basically repeating
what is happening inside the CatalogBalancer.

However I believe that the implementation is quite reasonable and strike
a balance between functionality added and extremelly clean design.