The first important fix is about the test. The final `$` is necessary to tell grep to match a new line. Without it, grep match the few catalogs that have been normally added in the test. (The catalog under /a and /a/b)
Then we simply fix the logic, adding a new empty file called
.cvmfscatalog into the repository. Earlier it was a file called as the
base_directory where the tar is ingested.... No idea why, but the
failure of the test I believe was shadowing the error.

Finally to make everything work I had to implement a method on the
DummyDirectory. Without the implementation it would call a method in a
base class that simply assert(false), it is not bad to implement that
method for the Dummy, otherwise we would have too re-structure a lot of
code for benefits that I fail to see.