Create a simple unit file to run DUCC loop as a daemon
The first step in this commit is to create a template file for the unit file, the only variable in the template file is the installation prefix used by CMAKE. This is necessary since we need to provide an absolute path to ExecStart in the UnitFile.

The second step is to configure the template and install the configured file into `/etc/systemd/system`.

Finally we documented how to use the generated unit file.