Draft for switching to raw abort() to the possible use of throw.
The switch is necessary since we want to create a library `libcvmfs_server.so` where a raw aborting is not desiderable. However we would like to keep the original behaviour in the binaries.
The first idea was to have a much simpler function that using pre-compilation condition would either simply `abort()` (for the binaries) or throw an exception string (for the library).
This approach would imply to find good descriptive string for all the abort in the code. Some abort do not have any message associate with (implying that we don't expect them to ever been called), other are preceded by a log (implying that we are facing a known error condition).
Overall it seems complex to find reasonable string messages for all the panics and to format those messages correctly.

The proposed approach will allow to automatically replace either the call to plain `abort(void)` with a call to `Panic(void)` or the call to `LogCvmfs(source, mask, format, ...)` to a call to `Panic(source, mask, format, ...)` leaving untouched the original semantic of the calls.

Downside of this approach is the problem in compilation phase annotate in the comment.
`Panic/3+` should be defined taking as first argument `const LogSource source` while now it is just `const int source` (it is correctly defined in the .cc file). Indeed including `logging.h` break the compilation process, this should be investigate and understood correctly