GetManifest now use cache, and add methods GetConfig and GetDiffIDs
This method is a slight refactor of the Image code. The main goal is to reach the GetDiffIDs functions with returns the DiffIDs of an image, those IDs are stored in the Config (hence the GetConfig).
Also we fixed an error on the use of pointers, now GetManifest takes the image as reference, hence can modify it, and this enable the cache in such a way that multiple calls to GetManifest results in only one call to the network