DownloadManager fix: switch hosts after redirect
When a host redirects to another failing URL, DownloadManager
incorrectly does not switch to next host. This is caused by
assuming that CURLINFO_EFFECTIVE_URL variable does not change
after a redirect, when it actually does.

Rewrite the 'do_switch' condition in SwitchHost() without the
help of CURLINFO_EFFECTIVE_URL variable.