make the gateway NOT hang when the receiver crash
the gateway relies on reading data from the receiver in order to
return controll to the main thread.

When everything goes right, the gateway reads the correct amount of
data from the pipe and exit.
When the receiver crash, for whatever reason, the data is not
present in the pipe. Moreover the pipe is not closed automatically
by the OS, since the File Descriptors are still open and managed by
the gateway itself.

This commit close the FDs on the gateway side, so that when the
receiver crash, no more FDs are open on the socket and the read
call can be unblocked by EOF.