Do not escape '@' in URI strings
RFC 3986 section 3.2.1 defines inline user information within a URI in
the form of "user:password@host".

The '@' character is currently escaped unconditionally when setting
CURLOPT_URL in DownloadManager::SetUrlOptions().  This causes any such
URIs to be treated as invalid and hence requests will eventually fail
with kFailBadUrl.

Fix by adding '@' to the list of characters that do not require URI
escaping.  The '@' character has no special meaning outside of the
user information field, and so this will not alter the interpretation
of any other part of the URI.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>