Allow ReportModification() to truncate recursion
Change the return type of the CatalogDiffTool::ReportModification()
callback to allow an implementation of ReportModification() to
indicate that further recursion is unnecessary.  This may be used to
save substantial amounts of work by optimising away unnecessary
recursions.

No recursion is performed after calling the ReportAddition() or
ReportRemoval() callbacks, and so the return type for those callbacks
does not need to be similarly modified.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>