Do not consume 100% CPU while waiting for upload jobs to complete
The current logic in SessionContext::UploadLoop() degenerates into a
CPU spinloop while waiting for the final upload jobs to complete.

Fix by using an end-of-queue dummy terminator job instead of an atomic
flag, thereby allowing the loop to sleep instead of spinning.

There seems to be no need to use the atomic compare-and-swap in
FinalizeDerived() to guarantee that pthread_join() is called exactly
once, since the surrounding control flow already appears to require
that FinalizeDerived() is called exactly once.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>