Ensure that catalog test history database filenames are unique
The gtest framework provides the concept of a "death test" to allow
for verification of code paths that are expected to terminate the
process.  With the "threadsafe" death test style, this is accomplished
by restarting the unit test binary as a subprocess with some
additional flags to cause the subprocess to run only the single test
in question.

Any initialization performed by the unit test binary will therefore be
repeated.  This causes problems for the catalog test history database,
which attempts to create a new SQLite database using a fixed filename
within the (temporary) test directory.  Since the filename is fixed,
repeated initialization will end up opening the existing database and
failing on the CREATE TABLE statements since the tables in question
already exist.

Fix by using a unique temporary filename for the history database.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>