Remove SessionContextBase::queue_was_flushed_
The queue_was_flushed_ FIFO seems to serve no useful purpose.  The way
it is implemented, it blocks the WaitForUpload() call until a
submitted object pack in flight is sent and submitted to the gateway.
However, that logic only starts as of the second WaitForUpload() call
because the queue_was_flushed_ is initially true.  Also, the upload
queue works on entire object packs - so as long as an object pack is
not full (and hence not dispatched), WaitForUpload() doesn't wait.

The barrier for the gateway uploader is actually not WaitForUpload()
but FinalizeSession().  Finalizing the session will end up in
SessionContextBase::Finalize, which looks at the currently open object
pack and closes out the tab itself.