Do not consume 100% CPU while waiting for S3 upload jobs to complete
The current logic in S3Uploader::MainCollectResults() is a pure CPU
spinloop that lasts for the duration of any S3 upload.  The call to
sched_yield() ensures that other threads are allowed to execute, but
produces no blocking behaviour that will prevent the mostly-idle
thread from consuming 100% of a CPU core.

Fix by extending the existing pipe-based pointer queue in
S3FanoutManager to add a queue of pointers to completed jobs that can
be retrieved by S3Uploader::MainCollectResults(), with a NULL pointer
used to request termination of the thread.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>