Include explicit compression mechanism within graft files
Files are marked as compressed or uncompressed at the level of
individual catalog entries, and the "cvmfs_swissknife graft" tool is
already capable of creating graft marker files for either compressed
or uncompressed versions of the file.  However, the information about
the compression mechanism is not stored within the graft marker file,
which makes it difficult to manage a repository containing a mixture
of compressed and uncompressed files.

Add a "compression=..." line to the graft marker files and ensure that
the compression method is propagated into the newly created catalog
entry.

For backwards compatibility, treat an absent "compression=..." line as
meaning that the compression mechanism should be inferred from the
sync mediator in use.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>