Allow upload_into_backend() to upload files during a gateway test
The "cvmfs_swissknife upload" command cannot be used to upload files
via a gateway.  Since integration tests always run with both the
gateway and publisher on the same machine, adapt the logic in
upload_into_backend() to perform the upload directly to the backend
storage even when a gateway is in use.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>