Handle zero-length uncompressed objects in object pack consumer
The current logic in ObjectPackConsumer::ConsumePayload() behaves
incorrectly in the case of a zero-length object at the end of the
buffer: the loop terminates prematurely since the end of the buffer
has been reached, but the method then returns kStateContinue since not
all objects have been processed.  This causes the receiver to
erroneously wait for further data that will never arrive.

Fix by ensuring that the loop continues to process a zero-length
object at the end of the buffer.

Note that a zero-length object can occur only when using uncompressed
objects, since even a zero-length object will have a non-zero length
after compression.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>