Do not truncate recursion for a modification above the lease path
Commit e9df405 ("Avoid recursing into paths outside the lease when
performing a merge") introduced a bug by omitting to consider the case
of a modification detected in a directory that is a parent of the
lease path.  Such a path is not within the lease path, but a child
path may still lie within the lease path.

Fix by continuing recursion if the current path is a parent of the
lease path, and add a test case that covers this scenario.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>