Allow nested catalog to be swapped even if already attached
There exist situations in which SwapNestedCatalog() may encounter a
catalog that has already been attached.  For example: if a directory
is converted into a nested catalog then any child catalogs will be
attached during the call to CreateNestedCatalog() on that directory.
A subsequent call to SwapNestedCatalog() for one of the child catalogs
will currently hit an assertion failure due to the old catalog having
already been attached.

Fix by detaching the existing catalog if found, after checking that
the catalog has not been modified.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>