Avoid diff recursion outside of the lease path during a catalog merge
The existing logic allows for recursion to be terminated early upon
encountering a reported modification that lies outside of the lease
path, in order to avoid wasting time comparing subdirectories that can
never be of interest.

Generalise this to terminate recursion as soon as any path that lies
outside of the lease path is encountered, even if such a path does not
contain a modification.

This generalisation requires centralising the concept of an ignored
path into the DiffRec() function.  For the sake of consistency, also
centralise the related concept of a reportable path, and test for
reportability prior to calling ReportModification() et al.

Extend the gateway integration tests to include modifications made for
all three possibilities: above, within, and outside the lease path.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>