Use systemctl to unmount filesystems when available
When systemctl is available, cvmfs_suid_helper will use it to mount
filesystems in order to prevent the cvmfs2 process from being
terminated unexpectedly by systemd (see CVM-1398 for details).

Generalise this logic to use systemctl for unmounting as well as
mounting, to avoid ending up in a situation in which the systemd mount
unit is still marked as active even though cvmfs_suid_helper has
unmounted the corresponding filesystem.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>