Ensure that filesystem is really mounted when using systemctl
On some versions of systemd (observed with systemd 248 on Fedora 34),
the systemd mount unit activity status seems not to track the mount
state.  In particular, the following sequence of events:

  - systemctl start cvmfs-repo.name.mount
  - umount /cvmfs/repo.name
  - systemctl start cvmfs-repo.name.mount

will fail to remount the filesystem, since systemd believe that the
mount unit is still in the active state after the manual "umount".

Work around this problem by using "systemctl restart" instead of
"systemctl start" in cvmfs_suid_helper.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>