Use "Cache-Control: no-cache" instead of bypassing system proxy
The get_item() shell function has an optional "noproxy" parameter.
The purpose of this parameter is unclear, but the most plausible
explanation seems to be as a means of ensuring that the download
fetches the current upstream file, ignoring any versions that may have
previously been cached in intermediate proxy servers.

This fails on any system where use of a proxy server is required for
all downloads, since direct origin server fetches are not permitted on
such systems.

Unconditionally add the HTTP/1.1 standard request header
"Cache-Control: no-cache" to force an origin server download, thereby
eliminating the need to bypass the system proxy.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>