Allow use of a proxy server for uploads to S3 storage
Allow a proxy to be specified via CVMFS_S3_PROXY in the s3.conf file.
This can be useful when the proxy is shared between several publishers
and a gateway machine since it allows the proxy to be configured to
cache uploaded files, thereby eliminating an extra round-trip when the
gateway needs to retrieve files that were uploaded by the publisher.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>