Detect explicit proxy errors
The existing logic attempts to guess whether an error represents an
origin server error or a proxy error using a heuristic based on the
status code alone.  If this heuristic incorrectly identifies a proxy
error as an origin server error then it may result in an unnecessary
I/O failure.  (The converse is not true: if the heuristic incorrectly
identifies an origin server error as a proxy error then it will merely
result in unnecessarily retried origin server requests.)

Check for the X-Squid-Error header as generated by Squid and the
Proxy-Status header as defined by the draft-ietf-httpbis-proxy-status
draft RFC, and reclassify a heuristically deduced origin server error
as a proxy error if these headers explicitly indicate that a proxy
error occurred.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>