Fix usage of opt_proxy_groups_current_burned_ to match documentation
The comment for opt_proxy_groups_current_burned_ defines it as the
"number of proxy servers that failed within current load-balance
group".  There is currently a confusing mixture of code paths that
treat this variable as a zero-based or one-based index.

Use this variable consistently as a zero-based index (i.e. with a
value of zero meaning that no proxies have yet failed), to reduce
confusion.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>