Proxy sharding (#2812)
* Allow ValidateProxyIpsUnlocked() to indicate potential proxy changes

With multiple concurrent proxies in use, the process of selecting a
proxy is potentially expensive.  Allow ValidateProxyIpsUnlocked() to
indicate whether or not the proxy list may have changed, to avoid the
need to always reselect a proxy after verifying that the DNS records
are up to date.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>

* Allow for proxy to be selected on a per-transfer basis

Allow for a proxy selection mechanism that can choose between any of
the currently non-burned proxies for a particular transfer.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>

* Provide a partial hash digest for use as an unsigned integer key

The existing digest algorithms satisfy the avalanche criteria and so
any partial subset (e.g. the first 32 bits of the digest result) may
sensibly be used as a key in contexts where collisions are unimportant
(e.g. for load-balancing).

Provide a helper method that extracts the first 32 bits of the digest
result as an unsigned 32-bit integer (in an endian-independent way).

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>

* Allow for requests to be sharded across multiple proxies

Implement proxy sharding via consistent hashing, with the proxy hash
keys calculated from the (DNS-resolved) URLs and the request hash keys
derived from the already-known content hash.

Handle the single-proxy case by building a map that contains only a
single entry, thereby allowing the new code paths to match the current
behaviour.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>

* Define CVMFS_PROXY_SHARD configuration option

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>

* Run proxy failover tests in both single-proxy and sharded modes

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>