Eliminate duplicate code in Publisher constructors
There are currently two versions of each of the Repository and
Publisher constructors in order to support the Publisher::Create()
method, which needs to construct a Publisher object before the
underlying repository files and directories have been created.  This
results in some duplicated logic.

Replace with a boolean "exists" parameter (defaulting to true) so that
there is no need to duplicate this code.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>