Use objects to represent server lock files
The ServerLockFile behaviour is currently transliterated directly from
the original shell scripts as a series of static methods with no
underlying objects, leading to code that is difficult to follow and
unsafe since there are multiple call sites that must manually obtain
the path.

Fix by converting the ServerLockFile class to represent the underlying
lock file, with the knowledge of the path encoded exactly once (during
the Publisher constructor).

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>