Separate repository lock and flag file logic
The "in_transaction.lock" file is a flag file whose validity extends
beyond the lifetime of a single process, and is conceptually unrelated
to the lock files such as "is_publishing.lock".

Split out the logic for flag files to a separate ServerFlagFile class
(with equivalent changes to the shell scripts) to make the code easier
to follow.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>