Clarify semantics of server lock file acquisition
The ServerLockFile::Acquire() method returns a boolean indicating
whether or not the lock was acquired.  Rename the method to TryLock()
to make this obvious at the point of use.  Add a method Lock() which
throws an exception if the lock cannot be obtained, and use this
within Publisher::Sync() since the code there clearly assumes that the
lock has been obtained.

Also rename Release() to Unlock() for consistency, and to move closer
to the C++11 standard Lockable method names lock(), try_lock() and
unlock().

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>