Remove concept of channel from History class
The concept of a numeric update channel predates the introduction of
named branches.  It is obsolete and unsupported: the command-line
tools provide no sensible way to handle channels, and there is no way
to configure a client to track a particular channel in its mountpoint.

Remove this last redundant trace of the concept of a channel, and
update the comments within the code to reflect that branches should be
used instead.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>