Introduce new timer interface
The current HighPerformanceTimer class only provides timing support
for an entire function and needs to be connected to a histogram object
in order to provide useful output.

This commit introduces a new timer interface, that provides
a simple way of timing a certain piece of code.

Also, a parent - child relationship is created automaticaly when a timer
is called during the run of another one, so that the cvmfs_talk output
can show a "partial flamegraph" for the code paths where timers were introduced.