Add average times profiling tool
This commit adds a profiling tool that provides user,
system, and total time statistics for a cvmfs benchmark.
Provided with a given benchmark script, this tool runs it for
a given number of times in all three cache scenarios and prints
the average times values, together with the ratios between
different cache scenarios.