Add tool for generating flamegraphs
Flamegraphs are a useful tool for profiling cvmfs. This commit adds a script
that can use different profilers to generate flamegraphs for a given
cvmfs repository. The flamegraphs can be generated during the run of a specific
benchmark or for a fixed amount of time.
This is still to be improved, since it can generate inaccurate results for some
multithreaded runs of cvmfs, and it also introduces a significant performance
overhead for long runs.