[cleanup] fix syntax highlighting in test scripts (#3508)
This is a purely decorative change, in the sense that it should have no effect on any functionality, but fixes syntax highlighting. Since the integration test scripts have no .sh suffix, most editors will display them as text files, in particular also the github web page renders them in this not-very-readable way (VSCode manages to infer that it's a script).

Adding a #!/bin/bash throughout seemed the easiest way of fixing this - it maybe slightly confusing because none of these scripts are executed (all of them are sourced and then cvmfs_run_test() is run by another script), but it provides enough of a hint to editors to make them more readable.

I also take the opportunity to harmonize file permissions on the main scripts - a few of them had +x set, I've removed this for all of them.