[gw] add regression test for lease db locking
This test reproduces the issue of lease end requests failing with

```
Lease end request - error reply: {"reason":"delete statement failed: database is locked","status":"error"}
```

as reported originally by LHCb (#3259).

The problem has been fixed in #3258 by liberally applying mutexes, but at the cost of performance regressions in other API operations, as reported by Matt (#3460). This patch supplements a regression test in order to be able to amend the original fix.

The test here handcrafts the API requests to the gateway instead of using transactions, which would work too, in order to avoid the overhead of running multiple publishers. It fairly consistently reproduces the error when testing a non-fixed gateway ( I tried to find a compromise between runtime and constency here. Sleeping between requests seems to had little effect)  and runs successfully with the fix.

Since it only gets and cancel leases, the test may not be exhaustive, but it's a start.