Fix ATLAS tightcache find test (#3560)
This test keeps failing with Input/Output errors as the repository grows.
It is likely that this is just the increased size of pinned catalogs that do not fit anymore in the purposefully small cache, but it should be debugged with a recent root hash. For now, use an older root hash that is known to work in this test to fix the test suite.