remove external package ipaddress
This package is a backport of the ipaddress module of the python standard library for python 2 and < 3.3.
Since there is a typo in version 1.0.22 on line 1106:

raise TypeError("%s and %s are not of the same version" (a, b))

it seems preferable to try to remove this external package altogether - there should be no platform anymore that needs it.

This implies that python 2 support is dropped ( which we still had through use of compatibility libraries).