Create scratch dir with consistent permissions
The scratch dir (typically under /var/spool/cvmfs/@fqrn@/scratch/current) gets initialised with mode 700, Publishing a transaction will recreate it with 755.
Aborting a transaction will reinit the directory with mode 700.  Since this directory is used in the overlay for /cvmfs/@fqrn@ the permissions for the top level dir
also seem to change, with potential for confussion. For consistency, this patch already initialises the scratch dir with mode 755.

This patch also fixes Publisher::CreateDirectoryAsOwner(const std::string &path, int mode), which ignored the mode parameter, hardcoding it to 700,