mkfs: do not overwrite existing manifest in s3 bucket
When using mkfs+s3 with a bucket already containing a repository, we will just overwrite the existing manifest, but it's very unlikely that's what the user actually wants to do. For local repositories on local storage this already fails due to the existence of the repository configuration on the local disk. This patch adds a check in mkfs. If one actually does want to overwrite the manifest, it's now necessary to delete the old one manually from the bucket.