Implement replacement for libacl acl_to_xattr()
Up to now, the upcoming swissknife_ingestsql tool uses libacl (from
"acl" project) to convert ACL metadata from textual to xattr form.

But libacl does not export appropriate function.
For now a workaround of patching a bundled libacl was employed. The
functionality works, but dependency bundling can create friction to
adoption by some Linux distributions.

The needed patch was sent to the mailing list of acl package, but in the
last 1.5 months, it hasn't received any reaction.
https://lists.nongnu.org/archive/html/acl-devel/2025-04/msg00000.html

Exactly the same technical change was considered in 2018, but nothing
was changed: https://mail.gnu.org/archive/html/acl-devel/2018-02/msg00004.html

This changeset does not remove libacl dependency from the build system,
this will have to be done in a separate commit.

This changeset includes both the implementation and the tests.
Tests are based on reference output which is incorporated into the test
source code. For most of the test cases, the new implementation returns
exactly the same binary output and return code as the old one.

The new unittests module has a compile time option (disabled by default)
to run both new implementation and libacl functions on the test cases
and to compare their results. To enable this:

- uncomment `#define COMPARE_TO_LIBACL` in cvmfs/acl.h
- add `${LIBACL_LIBRARY}` to CVMFS_UNITTESTS_LINK_LIBRARIES in test/unittests/CMakeLists.txt