ci: correct the diff fed to clang-tidy-diff
It used `git diff -U0 HEAD HEAD~`, which is the inversion of the last
commit, to run the lint on.

Checkout in github actions works in such way that the last commit in the
workcopy is the simulated merge of the target branch into the base
branch.

So we just flip the order of the arguments.

Co-authored-by: Valentin Volkl <valentin.volkl@cern.ch>