loader: Have premount check for -oro instead of !-orw to determine readonly (#3872)
When doing premount look for the "-oro" fuse option instead of a lack of
"-orw". Apparently both are being included. This was causing the mount
to be done as read-write instead of read-only, and on Ubuntu 24 inside a
docker container it ended up with a different error from the `futimes()`
system call that julia was apparently not checking for.

- Fixes #3868

---------

Co-authored-by: Valentin Volkl <valentin.volkl@cern.ch>