Allow builds without libfuse2 (#3879)
Since hotpatching requirements can make it difficult to remove libfuse2,
we decided to build cvmfs without libfuse2 support for the platforms
that will be added in cvmfs 2.13.1 (alma10, debian13 , fedora42,
ubuntu25.04). This is a first step towards dropping libfuse2 which will
remove quite a bit of complexity in the code.

The cvmfs-fuse3 package is kept separate for now, to minimize
disruptions.

This PR:

* introduces a `-DBUILD_LIBFUSE2=no` cmake option that allows to not
build the relevant libraries
* changes libfuse3 from an optional to a required dependency in cmake
* makes cvmfs-fuse3 a required dependency of cvmfs
* updates rpm packaging to drop libfuse2 for fedora >= 42 and rhel >=10
* updates debian packaging to drop libfuse2 for debian >= 13 and ubuntu
>= 25.04
  * Fixes #3866
* Cleans up the control file, removing comments regarding multi-arch
support, and improving the formatting
* In order not to do too much preprocessing, the control file is the
same for older and newer platform and now recommends libfuse2 instead of
requiring it.