server: remove code working around absense of jq
cvmfs_server's JSON utility code makes effort to detect absense of jq.
This on its own produces a side-effect of jq absense not showing up in
error logs.

If jq is absent, there were different code paths to handle the
task at hand, sometimes giving a warning ("it's your responsibility to
ensure JSON is valid").

Some of such fallback code paths simply cannot count as accomplishing
the task:

* update_repo_status() does nothing
* get_json_field() always provides empty result

This seems risky - it is surprising upon discovery, so it may produce
surprising results in operation. Hopefully jq is ubiquitous enough to
not need any workarounds.

There seems to be no testing for jq-less configuration, because there is
a shell syntax error, as can be seen in test 689. I only stumbled upon
this because of running tests in a different setting.

	Verifying integrity of test.cern.ch...
	Inspecting log of references
	Inspecting tag database
	[inspecting catalog] 0b9e8b479c3d6a36e0c1f5a7cfac3492a8feeb04 at /
	no problems found
	/usr/bin/cvmfs_server: line 2206: return: : numeric argument required
	*** setup cleanup handler

This change doesn't make it safe to operate cvmfs_server in jq-less
configuration, because cvmfs_server doesn't use strict mode (at least
`set -eo pipefail`), and running it without jq will not abort the
execution immediately upon failure to launch jq, which may cause damage.
Consumers of packaged cvmfs which properly declare runtime dependency on
jq are safe, though. If this change is not acceptable on its own, it can
wait until possibly cvmfs_server adopts necessary elements of strict
mode.