test/src/084-premounted: rely on run.sh for workdir cleaning
This test's code has an immediate issue that upon test failure, workdir
is still wiped, preventing immediate analysis.

test/run.sh already takes care of creation of workdir (the test function is
launched with it as current working dir), and similarly of removing of
it upon successful completion.

We still need a cleanup procedure to unmount our mount.
This commit prevents multiple executions of the cleanup routine.